+++ src/writeData.h
@@ -32,7 +32,7 @@ Header Guards
 
 struct dataStruct
 {
-	long int id;  // Line Number
+	long long int id;  // Line Number
 	double coordinates[3];     // column 1
 };
